tests refactorings, now it's 100% test coverage   - unit tests in 'src/' just use `core::*`   - integration tests in 'tests/' can use `std::*`   - move 'impl_generator' tests to 'tests/'   - add test for `#[derive(Debug)]` in 'impl_generator' tests
